fix: 经纬度 float 类型转换 & Formatter 支持表格 PipeTables

+++ app/Commands/Thread/EditThread.php
@@ -130,8 +130,8 @@ class EditThread
         if (isset($attributes['longitude']) && isset($attributes['latitude'])) {
             $this->assertCan($this->actor, 'edit', $thread);
 
-            $thread->longitude = Arr::get($this->data, 'attributes.longitude', 0);
-            $thread->latitude = Arr::get($this->data, 'attributes.latitude', 0);
+            $thread->longitude = (float) Arr::get($this->data, 'attributes.longitude', 0);
+            $thread->latitude = (float) Arr::get($this->data, 'attributes.latitude', 0);
             $thread->address = Arr::get($this->data, 'attributes.address', '');
             $thread->location = Arr::get($this->data, 'attributes.location', '');
         }
